Remove unused import and change HttpRequest parent
Following up on some review comments I changed the HttpRequest and
HttpResponse classes to subclass from aiohttp.ClientRequest and
aiohttp.ClientResponse respectively.

The lint check was catching an unused import in hdrs.py which I added a
skip to since it's importing names into the module that are used
elsewhere.